apps/kensan/kensan-ai/pyproject.toml
[project]
name = "kensan-ai"
version = "0.1.0"
description = "AI agents for Kensan learning management app"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "anyio>=4.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.6.0",
    "pyjwt>=2.8.0",
    "openai>=1.0.0",
    "pgvector>=0.2.0",
    "boto3>=1.26.0",
    "python-multipart>=0.0.6",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-asyncpg>=0.48b0",
    "opentelemetry-instrumentation-httpx>=0.48b0",
    "opentelemetry-exporter-otlp-proto-http>=1.27.0",
    "tavily-python>=0.5.0",
    "google-genai>=1.0.0",
    "google-adk>=1.0.0",
    "pyiceberg[s3fs,pyarrow]>=0.9.1,<0.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
kensan-profile-summarizer = "kensan_ai.batch.run_summarizer:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kensan_ai"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
